+++ doc/changelog.markdown
@@ -1,3 +1,5 @@
+  * Secondary edge interpretation changed to report actual relation
+  tags (Dag and Marius)
   * Generated inflections maintained in the database (Marius)
   * Information structure annotation (Anders)
   * Dependency alignment (Marius)